Addon-docs: Pass everything through babel & partial TS conversion
