+++ addons/docs/src/frameworks/common/index.ts
@@ -1,0 +1,1 @@
+export * from '../../lib/getPropDefs';
