+++ addons/docs/src/mdx/fixtures/decorators.mdx
