+++ addons/docs/src/mdx/fixtures/story-def-text-only.mdx
