+++ addons/docs/src/mdx/fixtures/story-function.mdx
