+++ addons/docs/src/mdx/fixtures/story-object.mdx
