+++ addons/docs/src/mdx/fixtures/story-references.mdx
