+++ addons/docs/src/mdx/fixtures/vanilla.mdx
