+++ addons/docs/src/typings.d.ts
@@ -1,1 +1,2 @@
 declare module '@mdx-js/react';
+declare module '@storybook/addon-docs/mdx-compiler-plugin';
