+++ addons/docs/vue/preset.js
@@ -1,1 +1,1 @@
-module.exports = require('../common/makePreset')('vue');
+module.exports = require('../dist/frameworks/common/makePreset').default('vue');
